feat(tracer): Register variables in trace
This commit is a first part of a two-part change; the foundation is laid
here to register variables during execution (debugging) of a Noir
program. The second part change will convert all types, not just Field,
as can be seen in the code from this first part.

The tests deleted in this commit will be restored then.