chore: keccak256 in Noir (#5316)
# Description

## Problem\*

The PR uses the keccak permutation blackbox function in order to
implement Keccak256 in Noir.


## Summary\*
The Noir implementation is added to the stdlib and results in circuit
size similar to the Keccak256 blackbox.


## Additional Context
The Keccak256 blackbox is not removed because my first experiments with
variable-size keccak are not conclusive yet.


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>