feat: let LSP always work in a Noir workspace if there's any (#5461)
# Description

## Problem

Resolves #5460

## Summary

Sorry for the many commits, the PR was stacked on top of other PRs.
Here's the first relevant commit:
[link](https://github.com/noir-lang/noir/pull/5461/commits/2d4d013e26474e13b133071f88bfc08011a5430b)

Before this PR, LSP "find all references" and "rename" worked like this:
1. Find what's in the cursor location. That could either be a
definition, or it could be a reference to a definition (or, well,
nothing)
2. If it's a reference, go to the definition (otherwise keep the
definition)
3. From the definition, find all references to that definition.

That works well, but that always happened in a single Package. So, if
you ran this in a Package A, but there's Package B that depends on
Package B, references in B weren't returned (they can't be seen from A!)

To solve this, a few things were done here:
1. First, the old LSP code always used the `Nargo.toml` file nearest to
a file. That means that it would never find workspace. Now we continue
searching a `Nargo.toml` upwards even if we find one, up until the LSP
root directory, and keep the last one (if we find anothe one it's likely
a workspace `Nargo.toml`). With this change it means that when opening
any file in a package, if that package is included in a workspace (and
that workspace is in the editor) we actually invoke "nargo check" on
each package in the workspace, not just in the package where the file
is. That's the first step in actually tracking everything in the
workspace.
2. Second, when searching for references, we search for references in
all packages. To do this we first find a definition in the current
package, then search references to that definition in all other
packages.

I believe that with this, the LSP features are now much more usable!
Searching a type, a struct member, an alias, a global, etc., anywhere,
will always yield the same results. This is useful, and necessary, if,
say, we want to rename a type or a member that's used across the
workspace: changing that only in some packages wouldn't be (that)
useful.

Here it is in action, using https://github.com/vlayer-xyz/monorepo to
try it out:


https://github.com/noir-lang/noir/assets/209371/015b6baf-2e44-4f17-a651-361a2ee93dd5

Here `StorageWithinBlock` is defined in a `lib` project which has no
dependencies to other packages in the workspace. However, other packages
depend on it, for example `get_storage`, and those references are found.

Note: the first commit is a refactor so it's easier to see where
references to specific things are tracked. With just `add_reference` it
was kind of hard to debug when something didn't work well for a specific
thing (for example globals). It also removes some redundancy because in
some cases the `is_self_type` boolean doesn't make sense.

## Additional Context

None.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>