feat: skip reading values immediately after it being written into an array (#5449)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

I noticed a couple of cases where we were writing into an array and then
immediately reading the same index from it. This seems like something
that we should be able to unwind.

Draft as I want to do more thinking on how `EnableSideEffects` can
interact with this.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.