chore(ci): add workflow to run `nargo check` on external repos (#5355)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR runs `nargo check` on external repositories in order to get
advanced warning of whether any changes we're making are likely to be
breaking changes. I've kept this CI as non-blocking as we can't make
these repos update atomically with ourselves.


## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.