feat(nargo): Default expression width field in `Nargo.toml` (#5505)
# Description

## Problem\*

Resolves #4961 

## Summary\*

This PR let's a developer override the default expression width target
with a field in the `Nargo.toml` manifest.

Due to the limitations of the `clap` macros I had to switch
`expression_width` on the `CompileOptions` struct to optional. I then
hardcoded a constant that specifies we want width 4 as Noir's
compilation default. We then settle on a target expression width by
checking the following:
1. Is there a user supplied expression width from the CLI flag? If there
is use that width.
2. Is there no user supplied width and an expression width specified in
the manifest? Use the expression width in the manifest.
3. Is there no user supplied width and no width specified in the
manifest? Use the hardcoded default value.

## Additional Context

## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>