chore: Noir version for Pedersen commitment and hash (#5431)
# Description

## Problem\*

Pedersen commitment and Pedersen Hash iin Noir

## Summary\*
When the constant input PR is merged and synchronised, these optimised
version of Pedersen commitments and Pedersen Hash in Noir should be in
parity with gates count of Pedersen commitment blackbox function.


## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.