fix: Mutability in the comptime interpreter (#5517)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/5504

## Summary\*

Mutable variables weren't wrapped in a mutable reference before like
they are in SSA - they are now.

## Additional Context

This is an alternate version from the last PR - I've cherry-picked the
last PR so this can be merged into master instead of another branch.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>