feat(tracer): Add function params to trace
The `variables` in `StackFrame` are now sorted before construction of the
frame. This is an unenforced invariant of the type. The sorting is thus
not needed anymore after the field is accessed (previously, done in `fn
register_variables`).

The new field of `StackFrame` -- `function_param_indexes` -- is a vector
of indexes in this sorted vector (`variables`). It lists which variables
are function parameters.

`stack_trace` is renamed to `stack_frames` for increased symmetry of the
code (it's of type `Vec<StackFrame>`)

Finally, instead of an empty vec[], actual args are passed to
`register_call`, which are built on top of the other changes mentioned.

One question that remains open is whether `register_value` needs to be
called in `convert_params_to_args_vec` or not. This is noted in a
comment and will be addressed at a later point.