chore: Remove `--use-legacy` and resolution code (#5248)
# Description

## Problem\*

## Summary\*

Made this PR as an experiment to see how much legacy code we can remove
today.

It looks like we can remove the majority of resolution code at the
`dc_crate` level, but we can't remove the NameResolver or TypeChecker
since these are used by the aztec macros crate. This is something we'll
have to watch out for when the elaborator is used by default - that code
injected by macros will still use the old resolution & type checking
code. I'm not aware of any compatibility errors when running both but it
is something to keep in mind.

Edit: Removed the NameResolver and TypeChecker in a separate PR:
https://github.com/noir-lang/noir/pull/5525

## Additional Context

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.