chore: Remove the remainder of legacy code (#5525)
# Description

## Problem\*

## Summary\*

Removes the remainder of legacy code: the Resolver and the TypeChecker.

## Additional Context

Removing this bit required updating aztec_macros a bit to use the
elaborator instead of the resolver/type checker when inserting the code,
hence why this was split into a separate PR. The change should be minor.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.