fix: Add `comptime` to trait impls and structs (#5511)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This is somewhat of a quick fix to the issue in the trait constraint PR
and in @asterite's comptime prefix operator PR.

The issue was that the comptime functions/blocks were using other
functions which weren't elaborated yet. So they saw an empty function
body. Just adding `comptime` to a function in an impl wasn't sufficient
since the compiler didn't look in every function within an impl to know
whether to add the impl to the list of comptime items. Moreover, it
wouldn't be possible to only add this `comptime` to a subset of
functions within an impl so I decided to allow `comptime` as a keyword
on an entire impl as well - and this is what decides whether or not to
add it to the comptime items.

I'm not really satisfied with this solution since it overloads
`comptime` to mean both "run this at compile-time (always)" and "_able
to be run_ at compile-time" but I don't have a better solution right
now. To some degree the separation is necessary since we need to
separate runtime functions that can be modified by comptime functions
from comptime functions that cannot be modified (because we've already
run them).

## Additional Context

Requires https://github.com/noir-lang/noir/pull/5517 to work

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.