feat: LSP inlay hints for let and global (#5510)
# Description

## Problem

Resolves #5513

## Summary

Shows type hints for `let` and `global`. Note that right now they will
always show for `let` and `global` that don't have an explicit type.
Eventually it would be nice to have a configuration to toggle them, but
the noise they introduce isn't that much so we could do that in a
separate PR (also paving the way for configuring other types of inlay
hints, like parameter names or names next to closing braces).


https://github.com/user-attachments/assets/4b743873-e1bf-430b-8632-4020ea53fbd8

## Additional Context

If you try it, hovering over types might not work well because of #5515,
so you could merge that branch into this one before testing things.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>