fix: Error on empty function bodies (#5519)
# Description

## Problem\*

Resolves an issue mentioned in this PR:
https://github.com/noir-lang/noir/pull/5511 And in @asterite's comptime
prefix operators PR.

## Summary\*

Previously, if a function was every (erroneously) allowed to be called
in a comptime context before it had been elaborated, we'd see just an
empty block as its body. I've changed this so we don't default function
bodies to an empty block anymore. Using an option we can now properly
error in this case as well.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>