feat: add comptime support for `modulus_*` compiler builtins (#5530)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR changes the function `std::compat::is_bn254` to run in a
`comptime` context. To do this I've had to implement a few compiler
builtins within the interpreter.

I've had to mangle `std::compat::is_bn254` a bit to get around
restrictions on when a function can be called in a `comptime` context
however.

This works around around the loop unrolling issue experienced in #4535.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.