fix: never panic in LSP inlay hints (#5534)
# Description

## Problem

I noticed that the inlay hint request would sometimes panic. This
happens if the request happens when a source file is being edited but
its cached version (in a FileMap) is not yet updated. Then, the editor
might pass a line range that's bigger than what's cached... producing a
panic.

In reality, the panic comes from the `codespan_lsp` crate: a function we
call returns `Result`, but there's an `unwrap()` inside it making it
panic instead of returning a Result.

## Summary

Here we don't depend on `codespan_lsp` for doing that, and instead copy
the logic but never panic.

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.