fix: revert PR #5449 (#5548)
# Description

## Problem\*

PR #5449 introduces a new issue in aztec protocol circuits

## Summary\*
Revert the PR


## Additional Context
The test 'sort_get_split_order_hints_desc_non_zero_split_counter_empty'
is failing with the PR.
It is not clear to me why the PR is making this to fail.
One corner case that we could have when you replace a load with the
stored value is:
- loose the index bound check (i.e `A[i]=y; x=A[i] +5; `becomes `x=y+5`,
whether or not i is out-of-bounds.
- loose the side-effect; if the store is under a side-effect, replacing
it uses by the stored value is not conditional to the enable-side-effect
value anymore.


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.