fix(ssa): More robust array deduplication check (#5547)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/5529

## Summary\*

In https://github.com/noir-lang/noir/pull/5287 we added a check to avoid
duplication of arrays. However, we can make this check more complete by
allowing us to check constant arrays based off of their contents when we
are inside of ACIR and also distinguishing between arrays and slices.

The program in the issue, for any size input array, is now the same
number of ACIR gates for both versions (using a local array variable and
an array fetched from a function).

## Additional Context


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.