fix: let unary traits work at comptime (#5507)
# Description

## Problem

Resolves #5492

## Summary

In #5423 we made unary traits work at runtime: this PR makes them work
at comptime.

## Additional Context

For some reason, in the test program, `_result` ends up being `()` when
I print it. It's the same if I implement `Add` for a type and use it a
comptime, so maybe it's expected or maybe that's not yet implemented?

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>