fix(frontend): Disallow signed numeric generics (#5572)
# Description

## Problem\*

Resolves #5552

## Summary\*

This temporarily disallows signed numeric generics as to prevent the
panic in the issue until they are fully supported. In general type-level
integers need to be updated to support numeric generics. It is deceiving
to not ban them until type-level integers are updated as otherwise users
can declare signed numeric generics but not actually use them as per the
linked issue.

For the code in the issue we now get this error:
<img width="697" alt="Screenshot 2024-07-19 at 5 41 13 PM"
src="https://github.com/user-attachments/assets/17979de7-8c82-40af-aa0f-a32f73992d48">


## Additional Context



## Documentation\*

Check one:
- [] No documentation needed.
- [ ] Documentation included in this PR.
- [X] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.