fix: Run macros within comptime contexts (#5576)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/5575

## Summary\*

I've given the interpreter a reference to the elaborator. When a macro
is called we setup that elaborator to elaborate in the middle of the
current function by saving any previous state and setting any state we
can. This is sufficient to resolve local variables, global variables,
generics, and some metadata like constrained-ness.

## Additional Context

To implement this I had to tell the elaborator not to call macros that
are within a comptime scope (wait for the interpreter to do it). This
lead to type errors in e.g. `assert(unquote!(...))` since the elaborator
would now see a `Quoted` type instead of the type it unquotes to. I've
put a band-aid over this by returning an unbound type variable in this
case. This usually works ok but in certain situations users may need a
type annotation when using macros within other comptime code. I think
this is a reasonable limitation.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.