feat: Implement `zeroed` in the interpreter (#5540)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Implements `zeroed` in the interpreter

## Additional Context

This is still a draft because I wanted to call attention to the
inability to actually implement `zeroed` when the types may still
sometimes not be known while interpretering. In this case we still don't
want to fail since these zeroed values aren't meant to be used anyway -
we'd rather return any dummy value. Here I've done it by adding
`Value::Zeroed` but I think we should perhaps consider changing `zeroed`
itself to return a dummy value of any type rather than one that is
guaranteed to be some zero equivalent.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.