feat(tracer): Fix steps before call
There was a problem with recording the correct step when there's a
function call.

This commit fixes the problem by switching to a newer version of
`runtime_tracing` (coordinated with the repo owners to mint this new
version) and our code now skips writing the step that points to the line
with the function declaration.

Inspecting the change in the tests you will notice that some values are
missing now. This is due to the the aforementioned dependency and if it
changes it will be due to a change there. I think it's fine as it is,
but opinion might vary.