chore: Switch `Value::TraitConstraint` to a resolved trait constraint (#5618)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Changes `Value::TraitConstraint` to hold a resolved trait constraint
rather than an unresolved one. This was the original intention of the
type but wasn't possible at the time since we didn't have access to the
elaborator to resolve it. Keeping it unresolved lead to issues with
hashing being different for constraints with different spans or with
some spans having `Some(trait_id)` and it being `None` for others. Now
they're all uniform.

## Additional Context

With this, the `derive` example now passes!

Putting it in the stdlib gives some odd errors for now though, so I'm
adding it as a test and will add it to the stdlib in a later PR.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.